divide functions in more files
